ARTEMIS-3261 Updating logic to use only replaceable records on compacting verification

+++ artemis-journal/src/main/java/org/apache/activemq/artemis/core/journal/Journal.java
@@ -105,7 +105,7 @@ public interface Journal extends ActiveMQComponent {
       appendAddRecord(id, recordType, EncoderPersister.getInstance(), record, sync, completionCallback);
    }
 
-   default void replaceableRecord(int recordType) {
+   default void replaceableRecord(byte recordType) {
    }
 
    void appendUpdateRecord(long id, byte recordType, byte[] record, boolean sync) throws Exception;
